requirements.txt
streamlit
numpy
opencv-python
matplotlib
Pillow
flask
scipy
scikit-learn
